Be more lax about token time limit
While testing locally against MediaWiki, I found on various occasions
this test failed, as the token was not completely in sync with my
local time (maybe an issue with my/MediaWiki's clock)

Allowing 1 second either side of the token removed all these problems.
I don't know enough about the oauth spec to know if this will be an
issue but it was causing me a lot of headaches during development at
the hackathon!